+++ test/com/onsoft/glasscat/install/utils/DefaultGpmPropsFactoryTest.ts
@@ -14,13 +14,7 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-import { TestSuite, Test } from "jec-juta";
-import { expect } from "chai";
-import { InstallDefaultGpmTaskProps } from "../../../../../../src/com/onsoft/glasscat/install/utils/InstallDefaultGpmTaskProps";
-import { GpmRef } from "../../../../../../src/com/onsoft/glasscat/install/utils/GpmRef";
-
-// utilities:
-//import * as utils from "../../../../../../utils/test-utils/utilities/InstallDefaultGpmTaskPropsTestUtils";
+import { TestSuite } from "jec-juta";
 
 @TestSuite({
   description: "Test the InstallDefaultGpmTaskProps class methods",